Allow compound regions and better background removal (#26)
Now allows for compound source and background regions.

Only subtracts off the intersection of any background regions from the source region (i.e. just the overlapping bits).

Tested with a number of overlapping rectangles and this seems to work.

Not sure how this will scale for larger numbers of source/background regions given that we have to loop over all of the source and background regions for each  DET1 exposure map iteration.