Observation class update
seqid= is now required when spawning the Observation() class.

Updated a lot of doctstrings and the example notebook for how to set various things.

Addresses #30 and now allows the evdir to be specified. See the notebook for an example.

Updated the notebook to show out out_path can be used.

Depricated set_seqid since that is now required (really, I just hid it from the user).