Fixes things when GOES changes
In the past, we could always use either GOES 15 or 17. GOES 17 had an issue and has been replaced by GOES 18.

Fixed this by now doing a SunPy Fido query over the time range of interest, figuring out what satellites are available, and then using the one with the highest number (assuming it'll always be the best in the case where you have both 16 and 18 or 15 and 17). Should (??) be future proof.